feat: implement choices chart tooltip
