+++ front_end/src/components/multiple_choice_chart_card/choice_checkbox.tsx
