+++ front_end/src/components/numeric_chard_card/numeric_cursor_item.tsx
